[Maps] split out DrawFilterControl and DrawControl (#95255)
* [Maps] split out DrawFilterControl and DrawControl

* clean up

* update i18n id

* give 'global_index_pattern_management_all' permission to functional test because new check blocks access without it

* revert last change
