+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/draw_control.tsx
@@ -12,20 +12,10 @@ import MapboxDraw from '@mapbox/mapbox-gl-draw';
 // @ts-expect-error
 import DrawRectangle from 'mapbox-gl-draw-rectangle-mode';
 import { Map as MbMap } from 'mapbox-gl';
-import { i18n } from '@kbn/i18n';
-import { Filter } from 'src/plugins/data/public';
-import { Feature, Polygon } from 'geojson';
-import { DRAW_TYPE, ES_GEO_FIELD_TYPE, ES_SPATIAL_RELATIONS } from '../../../../common/constants';
-import { DrawState } from '../../../../common/descriptor_types';
-import { DrawCircle, DrawCircleProperties } from './draw_circle';
-import {
-  createDistanceFilterWithMeta,
-  createSpatialFilterWithGeometry,
-  getBoundingBoxGeometry,
-  roundCoordinates,
-} from '../../../../common/elasticsearch_util';
+import { Feature } from 'geojson';
+import { DRAW_TYPE } from '../../../../common/constants';
+import { DrawCircle } from './draw_circle';
 import { DrawTooltip } from './draw_tooltip';
-import { getToasts } from '../../../kibana_services';
 
 const DRAW_RECTANGLE = 'draw_rectangle';
 const DRAW_CIRCLE = 'draw_circle';
@@ -35,10 +25,8 @@ mbDrawModes[DRAW_RECTANGLE] = DrawRectangle;
 mbDrawModes[DRAW_CIRCLE] = DrawCircle;
 
 export interface Props {
-  addFilters: (filters: Filter[], actionId: string) => Promise<void>;
-  disableDrawState: () => void;
-  drawState?: DrawState;
-  isDrawingFilter: boolean;
+  drawType?: DRAW_TYPE;
+  onDraw: (event: { features: Feature[] }) => void;
   mbMap: MbMap;
 }
 
@@ -70,86 +58,12 @@ export class DrawControl extends Component<Props, {}> {
       return;
     }
 
-    if (this.props.isDrawingFilter) {
+    if (this.props.drawType) {
       this._updateDrawControl();
     } else {
       this._removeDrawControl();
     }
   }, 0);
-
-  _onDraw = async (e: { features: Feature[] }) => {
-    if (
-      !e.features.length ||
-      !this.props.drawState ||
-      !this.props.drawState.geoFieldName ||
-      !this.props.drawState.indexPatternId
-    ) {
-      return;
-    }
-
-    let filter: Filter | undefined;
-    if (this.props.drawState.drawType === DRAW_TYPE.DISTANCE) {
-      const circle = e.features[0] as Feature & { properties: DrawCircleProperties };
-      const distanceKm = _.round(
-        circle.properties.radiusKm,
-        circle.properties.radiusKm > 10 ? 0 : 2
-      );
-      // Only include as much precision as needed for distance
-      let precision = 2;
-      if (distanceKm <= 1) {
-        precision = 5;
-      } else if (distanceKm <= 10) {
-        precision = 4;
-      } else if (distanceKm <= 100) {
-        precision = 3;
-      }
-      filter = createDistanceFilterWithMeta({
-        alias: this.props.drawState.filterLabel ? this.props.drawState.filterLabel : '',
-        distanceKm,
-        geoFieldName: this.props.drawState.geoFieldName,
-        indexPatternId: this.props.drawState.indexPatternId,
-        point: [
-          _.round(circle.properties.center[0], precision),
-          _.round(circle.properties.center[1], precision),
-        ],
-      });
-    } else {
-      const geometry = e.features[0].geometry as Polygon;
-      // MapboxDraw returns coordinates with 12 decimals. Round to a more reasonable number
-      roundCoordinates(geometry.coordinates);
-
-      filter = createSpatialFilterWithGeometry({
-        geometry:
-          this.props.drawState.drawType === DRAW_TYPE.BOUNDS
-            ? getBoundingBoxGeometry(geometry)
-            : geometry,
-        indexPatternId: this.props.drawState.indexPatternId,
-        geoFieldName: this.props.drawState.geoFieldName,
-        geoFieldType: this.props.drawState.geoFieldType
-          ? this.props.drawState.geoFieldType
-          : ES_GEO_FIELD_TYPE.GEO_POINT,
-        geometryLabel: this.props.drawState.geometryLabel ? this.props.drawState.geometryLabel : '',
-        relation: this.props.drawState.relation
-          ? this.props.drawState.relation
-          : ES_SPATIAL_RELATIONS.INTERSECTS,
-      });
-    }
-
-    try {
-      await this.props.addFilters([filter!], this.props.drawState.actionId);
-    } catch (error) {
-      getToasts().addWarning(
-        i18n.translate('xpack.maps.drawControl.unableToCreatFilter', {
-          defaultMessage: `Unable to create filter, error: '{errorMsg}'.`,
-          values: {
-            errorMsg: error.message,
-          },
-        })
-      );
-    } finally {
-      this.props.disableDrawState();
-    }
-  };
 
   _removeDrawControl() {
     if (!this._mbDrawControlAdded) {
@@ -157,13 +71,13 @@ export class DrawControl extends Component<Props, {}> {
     }
 
     this.props.mbMap.getCanvas().style.cursor = '';
-    this.props.mbMap.off('draw.create', this._onDraw);
+    this.props.mbMap.off('draw.create', this.props.onDraw);
     this.props.mbMap.removeControl(this._mbDrawControl);
     this._mbDrawControlAdded = false;
   }
 
   _updateDrawControl() {
-    if (!this.props.drawState) {
+    if (!this.props.drawType) {
       return;
     }
 
@@ -171,27 +85,27 @@ export class DrawControl extends Component<Props, {}> {
       this.props.mbMap.addControl(this._mbDrawControl);
       this._mbDrawControlAdded = true;
       this.props.mbMap.getCanvas().style.cursor = 'crosshair';
-      this.props.mbMap.on('draw.create', this._onDraw);
+      this.props.mbMap.on('draw.create', this.props.onDraw);
     }
 
     const drawMode = this._mbDrawControl.getMode();
-    if (drawMode !== DRAW_RECTANGLE && this.props.drawState.drawType === DRAW_TYPE.BOUNDS) {
+    if (drawMode !== DRAW_RECTANGLE && this.props.drawType === DRAW_TYPE.BOUNDS) {
       this._mbDrawControl.changeMode(DRAW_RECTANGLE);
-    } else if (drawMode !== DRAW_CIRCLE && this.props.drawState.drawType === DRAW_TYPE.DISTANCE) {
+    } else if (drawMode !== DRAW_CIRCLE && this.props.drawType === DRAW_TYPE.DISTANCE) {
       this._mbDrawControl.changeMode(DRAW_CIRCLE);
     } else if (
       drawMode !== this._mbDrawControl.modes.DRAW_POLYGON &&
-      this.props.drawState.drawType === DRAW_TYPE.POLYGON
+      this.props.drawType === DRAW_TYPE.POLYGON
     ) {
       this._mbDrawControl.changeMode(this._mbDrawControl.modes.DRAW_POLYGON);
     }
   }
 
   render() {
-    if (!this.props.isDrawingFilter || !this.props.drawState) {
+    if (!this.props.drawType) {
       return null;
     }
 
-    return <DrawTooltip mbMap={this.props.mbMap} drawState={this.props.drawState} />;
+    return <DrawTooltip mbMap={this.props.mbMap} drawType={this.props.drawType} />;
   }
 }
