+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/index.ts
@@ -5,28 +5,4 @@
  * 2.0.
  */
 
-import { AnyAction } from 'redux';
-import { ThunkDispatch } from 'redux-thunk';
-import { connect } from 'react-redux';
-import { DrawControl } from './draw_control';
-import { updateDrawState } from '../../../actions';
-import { getDrawState, isDrawingFilter } from '../../../selectors/map_selectors';
-import { MapStoreState } from '../../../reducers/store';
-
-function mapStateToProps(state: MapStoreState) {
-  return {
-    isDrawingFilter: isDrawingFilter(state),
-    drawState: getDrawState(state),
-  };
-}
-
-function mapDispatchToProps(dispatch: ThunkDispatch<MapStoreState, void, AnyAction>) {
-  return {
-    disableDrawState() {
-      dispatch(updateDrawState(null));
-    },
-  };
-}
-
-const connected = connect(mapStateToProps, mapDispatchToProps)(DrawControl);
-export { connected as DrawControl };
+export { DrawFilterControl } from './draw_filter_control';
