+++ x-pack/plugins/maps/public/connected_components/mb_map/mb_map.tsx
@@ -17,7 +17,7 @@ import sprites2 from '@elastic/maki/dist/sprite@2.png';
 import { Adapters } from 'src/plugins/inspector/public';
 import { Filter } from 'src/plugins/data/public';
 import { ActionExecutionContext, Action } from 'src/plugins/ui_actions/public';
-import { DrawControl } from './draw_control';
+import { DrawFilterControl } from './draw_control';
 import { ScaleControl } from './scale_control';
 // @ts-expect-error
 import { TooltipControl } from './tooltip_control';
@@ -418,7 +418,7 @@ export class MBMap extends Component<Props, State> {
     let scaleControl;
     if (this.state.mbMap) {
       drawControl = this.props.addFilters ? (
-        <DrawControl mbMap={this.state.mbMap} addFilters={this.props.addFilters} />
+        <DrawFilterControl mbMap={this.state.mbMap} addFilters={this.props.addFilters} />
       ) : null;
       tooltipControl = !this.props.settings.disableTooltipControl ? (
         <TooltipControl
